requirements.txt
adal==1.2.7
argcomplete==2.1.2
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.29.4
azure-graphrbac==0.61.1
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.2.3
azure-mgmt-network==21.0.1
azure-mgmt-resource==22.0.0
azure-mgmt-storage==21.0.0
azure-storage-blob==12.18.1
azureml==0.2.7
azureml-core==1.53.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contextlib2==21.6.0
cryptography==41.0.4
distlib==0.3.7
docker==6.1.3
filelock==3.12.4
humanfriendly==10.0
idna==3.4
isodate==0.6.1
jeepney==0.8.0
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.2
knack==0.10.1
msal==1.24.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.26.0
oauthlib==3.2.2
packaging==23.0
pandas==2.1.0
paramiko==3.3.1
pathspec==0.11.2
pkginfo==1.9.6
platformdirs==3.10.0
portalocker==2.8.2
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.0
scipy==1.11.2
SecretStorage==3.3.3
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.5
websocket-client==1.6.3